// src/enums/State.enum.ts
export enum State {
  base = 'base',
  primary = 'primary',
  secondary = 'secondary',
  danger = 'danger',
  info = 'info',
  success = 'success',
  warning = 'warning',
}

export enum InputState {
  base = 'base',
  danger = 'danger',
  info = 'info',
  success = 'success',
  warning = 'warning',
}